EDA with visualisation
Exploratory data analysis of the SpaceX data that is supposed to guide us through our choice of variables. But domain knowledge is missing from these data to reach meaningful conclusions. For instance, VASB seems to be abandoned after June 2019, and we can only speculate as to why (and we'd be wrong, since launches are still planned for VASB/VSFB).
The python file is the short version of the Jupyter notebook, and the csv file is the last version of the dataset, with one-hot-encoding.